style: move the function tha creates the mock api to a new file
